attempt to fix ticket:330, to be tested
git-svn-id: https://mary.opendfki.de/repos/trunk@2660 953a6561-930b-0410-b2a6-db37d1b2ae63

+++ java/marytts/tools/install/InstallerGUI.java
@@ -535,7 +535,7 @@ public class InstallerGUI extends javax.swing.JFrame implements VoiceUpdateListe
     
     public void installSelectedLanguagesAndVoices()
     {
-        int downloadSize = 0;
+        long downloadSize = 0;
         List<ComponentDescription> toInstall = getComponentsSelectedForInstallation();
         if (toInstall.size() == 0) {
             JOptionPane.showMessageDialog(this, "You have not selected any installable components");
