+++ java/marytts/util/MaryUtils.java
@@ -987,13 +987,13 @@ public class MaryUtils {
         out.println(commentChar+" #########################################################################");
     }
 
-    public static String toHumanReadableSize(int byteCount) {
+    public static String toHumanReadableSize(long byteCount) {
         if (byteCount >= 10*1024*1024) {
             return (byteCount/(1024*1024))+"MB";
         } else if (byteCount >= 10*1024) {
             return (byteCount/1024)+"kB";
         } else {
-            return Integer.toString(byteCount);
+            return Long.toString(byteCount);
         }
     }
